+++ src/Components/OATPropertyEditor/Internal/EditorJsonTab/Internal/JSONEditor/JSONEditor.types.ts
@@ -1,9 +1,9 @@
 import { IStyle, IStyleFunctionOrObject } from '@fluentui/react';
-import { Theme } from '../../../../Models/Constants/Enums';
-import { IExtendedTheme } from '../../../../Theming/Theme.types';
+import { Theme } from '../../../../../../Models/Constants/Enums';
+import { IExtendedTheme } from '../../../../../../Theming/Theme.types';
 
 export interface IJSONEditorProps {
-    selectedTheme?: Theme;
+    selectedTheme: Theme;
     /**
      * Call to provide customized styling that will layer on top of the variant rules.
      */
